Extract StaticConfigTrait.
Extract a trait for the config(),drop(),configured() duck type used in
a number of classes. Implement the trait in Cache. This also removes the
Closure factory function. It will not be usuable by Email, and I think
it provides questionable value right now. If it becomes useful in the
future it can always be re-added.
